Renamed pnm files
